--- events.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "main.hpp"
-
-void dziekan_event (GameData& gameData);
-
-void zaionc_event (GameData& gameData);
