+++ rooms/windaPietro.cpp
@@ -4,7 +4,7 @@
 
 Room windaPietro (GameData& gameData)
 {
-	std::cout << "Winda nie działa, idziesz schodami."; 
+	std::cout << "Winda nie działa, trzeba iść schodami."; 
 	wait_s(2);
 	std::cout << std::endl;
